+++ R/Line.R
@@ -280,6 +280,15 @@ Line <- R6Class(
       Line$new(H, M, extendH, extendM)
     },
 
+    #' @description Parallel to the reference line passing through a given point.
+    #' @param M a point
+    #' @return A \code{Line} object.
+    parallel = function(M){
+      A <- private[[".A"]]; B <- private[[".B"]]
+      A_B <- B - A
+      Line$new(M, M + A_B)
+    },
+
     #' @description Orthogonal projection of a point to the reference line.
     #' @param M a point
     #' @return A point.
@@ -288,6 +297,14 @@ Line <- R6Class(
       A_B <- B - A
       v <- c(-A_B[2L], A_B[1L])
       .LineLineIntersection(A, B, M, M+v)
+    },
+
+    #' @description Distance from a point to the reference line.
+    #' @param M a point
+    #' @return A positive number.
+    distance = function(M){
+      P <- self$projection(M)
+      .distance(M, P)
     },
 
     #' @description Reflection of a point with respect to the reference line.