+++ NEWS.md
@@ -1,3 +1,22 @@
+# PlaneGeometry 1.1.0 (2020-02-24)
+
+* Bug fixed: the function `inversionSwappingTwoCircles` did not work for all cases.
+
+* New functions: `inversionFromCircle`, `midCircles`, `CircleAB`.
+
+* New methods for `Circle` class: `tangentsThroughExternalPoint`, 
+`isOrthogonal`, `angle`.
+
+* New methods for `Triangle` class: `symmedialTriangle`, `symmedianPoint`, 
+`BrocardCircle`, `BrocardPoints`, `LemoineCircleI/II/III`, `LemoineTriangle`, 
+`ParryCircle`, `SteinerEllipse`, `SteinerInellipse`, `pointToTrilinear`, 
+`pedalTriangle`, `CevianTriangle`.
+
+* New methods for `Line` class: `distance`, `parallel`.
+
+* New methods for `Mobius` class: `power`, `gpower`.
+
+
 # PlaneGeometry 1.0.0 (2020-02-01)
 
 First release.